+++ api_v3/services/UserService.php
@@ -38,70 +38,10 @@ class UserService extends KalturaBaseUserService
 		{
 			$user->isAdmin = true;
 		}
-
-		$this->validateUserNames($user);
+		$names = array('firstName', 'lastName', 'fullName', 'screenName');
+		$this->validateNames($user ,$names);
 		$lockKey = "user_add_" . $this->getPartnerId() . $user->id;
 		return kLock::runLocked($lockKey, array($this, 'adduserImpl'), array($user));
-	}
-	
-	function addUserImpl(KalturaUser $user)
-	{
-		/* @var $dbUser kuser */
-		$dbUser = $user->toInsertableObject();
-		$dbUser->setPartnerId($this->getPartnerId());
-		try {
-			$checkPasswordStructure = isset($user->password) ? true : false;
-			$dbUser = kuserPeer::addUser($dbUser, $user->password, $checkPasswordStructure);
-		}
-		
-		catch (kUserException $e) {
-			$code = $e->getCode();
-			if ($code == kUserException::USER_ALREADY_EXISTS) {
-				throw new KalturaAPIException(KalturaErrors::DUPLICATE_USER_BY_ID, $user->id); //backward compatibility
-			}
-			if ($code == kUserException::LOGIN_ID_ALREADY_USED) {
-				throw new KalturaAPIException(KalturaErrors::DUPLICATE_USER_BY_LOGIN_ID, $user->email); //backward compatibility
-			}
-			else if ($code == kUserException::USER_ID_MISSING) {
-				throw new KalturaAPIException(KalturaErrors::PROPERTY_VALIDATION_CANNOT_BE_NULL, $user->getFormattedPropertyNameWithClassName('id'));
-			}
-			else if ($code == kUserException::INVALID_EMAIL) {
-				throw new KalturaAPIException(KalturaErrors::INVALID_FIELD_VALUE, 'email');
-			}
-			else if ($code == kUserException::INVALID_PARTNER) {
-				throw new KalturaAPIException(KalturaErrors::UNKNOWN_PARTNER_ID);
-			}
-			else if ($code == kUserException::ADMIN_LOGIN_USERS_QUOTA_EXCEEDED) {
-				throw new KalturaAPIException(KalturaErrors::ADMIN_LOGIN_USERS_QUOTA_EXCEEDED);
-			}
-			else if ($code == kUserException::PASSWORD_STRUCTURE_INVALID) {
-				$partner = $dbUser->getPartner();
-				$invalidPasswordStructureMessage='';
-				if($partner && $partner->getInvalidPasswordStructureMessage())
-					$invalidPasswordStructureMessage = $partner->getInvalidPasswordStructureMessage();
-				throw new KalturaAPIException(KalturaErrors::PASSWORD_STRUCTURE_INVALID,$invalidPasswordStructureMessage);
-			}
-			throw $e;			
-		}
-		catch (kPermissionException $e)
-		{
-			$code = $e->getCode();
-			if ($code == kPermissionException::ROLE_ID_MISSING) {
-				throw new KalturaAPIException(KalturaErrors::ROLE_ID_MISSING);
-			}
-			if ($code == kPermissionException::ONLY_ONE_ROLE_PER_USER_ALLOWED) {
-				throw new KalturaAPIException(KalturaErrors::ONLY_ONE_ROLE_PER_USER_ALLOWED);
-			}
-			else if ($code == kPermissionException::USER_ROLE_NOT_FOUND) {
-				throw new KalturaAPIException(KalturaErrors::USER_ROLE_NOT_FOUND);
-			}
-			throw $e;
-		}
-
-		$newUser = new KalturaUser();
-		$newUser->fromObject($dbUser, $this->getResponseProfile());
-		
-		return $newUser;
 	}
 
 	/**
@@ -129,7 +69,8 @@ class UserService extends KalturaBaseUserService
 			throw new KalturaAPIException(KalturaErrors::CANNOT_SET_ROOT_ADMIN_AS_NO_ADMIN);
 		}
 
-		$this->validateUserNames($user);
+		$names = array('firstName', 'lastName', 'fullName', 'screenName');
+		$this->validateNames($user, $names);
 
 		// update user
 		try
@@ -701,19 +642,4 @@ class UserService extends KalturaBaseUserService
 
 		return $this->dumpFile($file_path, 'text/csv');
 	}
-
-
-	/**
-	 * @param KalturaUser $user The user parameters to validate
-	 */
-	protected function validateUserNames(KalturaUser $user)
-	{
-		$names = array('firstName', 'lastName', 'fullName', 'screenName');
-		foreach ($names as $name)
-		{
-			if (!is_null($user->$name) && strpos($user->$name, kuser::URL_PATTERN) !== false)
-				throw new KalturaAPIException(KalturaErrors::INVALID_FIELD_VALUE, $name);
-		}
-	}
-
 }