+++ plugins/group/esearch/api/types/items/KalturaESearchGroupBaseItem.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * @package plugins.group
+ * @subpackage api.objects
+ */
+abstract class KalturaESearchGroupBaseItem extends KalturaESearchBaseItem
+{
+
+}