+++ plugins/virus_scan/lib/batch/kVirusScanFlowManager.php
@@ -100,7 +100,14 @@ class kVirusScanFlowManager implements kBatchJobStatusEventConsumer, kObjectAdde
 		if (!class_exists('kCurrentContext') || !kCurrentContext::isApiV3Context())
 			return false;
 		if($object instanceof asset)
+		{
+			$entry = $object->getentry();
+			if ($entry && $entry->getFlowType() == EntryFlowType::IMPORT_FOR_CLIP_CONCAT )
+			{
+				return false;
+			}
 			return true;
+		}
 		if($object instanceof FileSync)
 			return true;